add separate brand and category data shapes for groq
This allows us to query for specific types, and
to find the relations between those types.

Groq seems to require specific keywords to figure this out. So each category
must have a `_id` field that matches the `_ref` field of the type
referring to it. It also requires `_type` of `reference`, I believe.